After we get a good power signal, always wait about 10ms before proceeding.
On boot, some laptops with certain cards in them sometimes fail on
boot, but if the card is inserted after boot it works.  Experiments
show that small delays here makes things more reliable.  It is
believed that some combinations need a little more time before the
power on the card is really stable enough to be reliable once the
power is stable in the bridge.


git-svn-id: http://svn.freebsd.org/base/head@223218 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f